feat(platform): integrate coaching page filters
