+++ packages/translations/src/lib/dictionaries/en.ts
@@ -3,6 +3,7 @@ import { Home_EN } from '../pages/home/home-en';
 import { Login_EN } from '../pages/login/login-en';
 import { SSO_EN } from '../pages/sso/sso-en';
 import { Offers_EN } from '../pages/offers/offers-en';
+import { Coaching_EN } from '../pages/coaching/coaching-en';
 
 
 export const EN: TDictionary = {
@@ -558,6 +559,7 @@ export const EN: TDictionary = {
     login: Login_EN,
     sso: SSO_EN,
     offers: Offers_EN,
+    coaching: Coaching_EN,
     auth: {
       errorPage: {
         title: 'Oops! There was a problem signing you in.',
