+++ packages/translations/src/lib/pages/coaching/coaching-de.ts
@@ -1,0 +1,6 @@
+import { TDictionary } from '../../dictionaries/base';
+
+export const Coaching_DE: TDictionary['pages']['coaching'] = {
+  filterByTopic: 'Suche nach Thema / Fähigkeit',
+  chooseCategory: 'Was ist Ihr Ziel?',
+};
