+++ packages/translations/src/lib/pages/coaching/coaching-en.ts
@@ -1,0 +1,6 @@
+import { TDictionary } from "../../dictionaries/base";
+
+export const Coaching_EN: TDictionary["pages"]["coaching"] = {
+  filterByTopic: "Search by Topic / Skill",
+  chooseCategory: "What's your goal?",
+}
